[agent] Add firstrun adoption tests
Adds adoption_test.go with 4 black-box tests covering PromptAdoption:
NoCandidates_NoOutput, UserAccepts_Adopts, UserDeclines_SkipsAdoption,
and IgnoresConfigMode (spec §8).
